+++ packages/core/theme/index.d.ts
@@ -1,40 +1,49 @@
-import * as React from 'react';
+declare module '@atlaskit/theme' {
+  import * as React from 'react';
 
-export const borderRadius: () => number;
-export const codeFontFamily: () => string;
-export const noFocusRing: () => string;
-export const focusRing: (color: string, outlineWidth: number) => string;
-export const colors: Record<string, string>;
-export const elevation: any;
-export const fontFamily: any;
-export const fontSize: any;
-export const gridSize: any;
-export const layers: Record<string, () => number>;
-export const math: any;
-export const Theme: React.ComponentType<any>;
-export const themed: any;
-export const typography: any;
+  export const borderRadius: () => number;
+  export const codeFontFamily: () => string;
+  export const noFocusRing: () => string;
+  export const focusRing: (color: string, outlineWidth: number) => string;
+  export const colors: Record<string, string>;
+  export const elevation: any;
+  export const fontFamily: any;
+  export const fontSize: any;
+  export const gridSize: any;
+  export const layers: Record<string, () => number>;
+  export const math: any;
+  export const themed: any;
+  export const typography: any;
 
-export type ThemeProp = <ThemeTokens, ThemeProps>(
-  themeFn: (ThemeProps: ThemeProps) => ThemeTokens,
-  themeProps: ThemeProps,
-) => ThemeTokens;
+  export class AtlaskitThemeProvider<ThemeProps> extends React.Component<
+    ThemeProps
+  > {}
 
-export class AtlaskitThemeProvider extends React.Component<ThemeProps> {}
+  export type ThemeProp<ThemeTokens, ThemeProps> = (
+    themeFn: (ThemeProps: ThemeProps) => ThemeTokens,
+    themeProps: ThemeProps,
+  ) => ThemeTokens;
 
-const Theme: {
-  Consumer: React.ComponentType<
-    ThemeProps & { children: (tokens: ThemeTokens) => React.ReactElement }
-  >;
-  Provider: React.ComponentType<{
-    value?: ThemeProp<ThemeTokens, ThemeProps>;
-    children?: React.ReactNode;
-  }>;
-};
+  export interface Theme<ThemeTokens, ThemeProps> {
+    Consumer: React.ComponentType<
+      ThemeProps & {
+        children: (tokens: ThemeTokens) => React.ReactElement<ThemeProps>;
+      }
+    >;
+    Provider: React.ComponentType<{
+      value?: ThemeProp<ThemeTokens, ThemeProps>;
+      children?: React.ReactNode;
+    }>;
+  }
 
-export default Theme;
-export const GlobalTheme: Theme;
+  export type GlobalThemeTokens = {
+    mode: 'dark' | 'light';
+  };
 
-export const createTheme: <ThemeTokens, ThemeProps>(
-  theme: (props: ThemeProps) => ThemeTokens,
-) => Theme;
+  const GlobalTheme: Theme<GlobalThemeTokens, any>;
+  export default GlobalTheme;
+
+  export const createTheme: <ThemeTokens, ThemeProps>(
+    theme: (props: ThemeProps) => ThemeTokens,
+  ) => Theme<ThemeTokens, ThemeProps>;
+}