+++ packages/editor/editor-core/src/plugins/media/ui/ResizableMediaSingle/styled.tsx
@@ -21,5 +21,9 @@ export const Wrapper: React.ComponentClass<
     content: '';
     display: block;
     padding-bottom: ${p => (p.height / p.width) * 100}%;
+
+    /* Fixes extra padding problem in Firefox */
+    font-size: 0;
+    line-height: 0;
   }
 `;