+++ packages/elements/status/example-helpers/ManagedStatusPicker.tsx
@@ -1,6 +1,5 @@
 import * as React from 'react';
-import { Color as ColorType } from '../src/components/Status';
-import { StatusPicker } from '../src/components/StatusPicker';
+import { StatusPicker, ColorType } from '../src/picker';
 
 export interface Props {
   initialSelectedColor: ColorType;