+++ packages/media/media-editor/src/react/editorView/toolbar/buttons/shapeButton.tsx
@@ -1,14 +1,15 @@
 import * as React from 'react';
 import { Component } from 'react';
+import { injectIntl, InjectedIntlProps } from 'react-intl';
 import ChevronDownIcon from '@atlaskit/icon/glyph/chevron-down';
 import Button from '@atlaskit/button';
 import Tooltip from '@atlaskit/tooltip';
-import { injectIntl, InjectedIntlProps } from 'react-intl';
+import { messages } from '@atlaskit/media-ui';
+import { colors } from '@atlaskit/theme';
 import { toolIcons } from './toolButton';
 import { Tool } from '../../../../common';
 import { shapeTools } from '../popups/shapePopup';
 import { DropdownLeftIconWrapper, DropdownRightIconWrapper } from './styles';
-import { messages } from '@atlaskit/media-ui';
 
 export interface ShapeButtonProps {
   readonly activeShape: Tool;
@@ -26,17 +27,22 @@ export class ShapeButton extends Component<
       activeShape,
       intl: { formatMessage },
     } = this.props;
+    const iconPrimaryColor = isActive ? colors.N0 : undefined;
     const isShapeTool = shapeTools.indexOf(activeShape) > -1;
     const Icon = toolIcons[isShapeTool ? activeShape : shapeTools[0]];
 
     const iconBefore = (
       <DropdownLeftIconWrapper>
-        <Icon label={activeShape} size="medium" />
+        <Icon
+          label={activeShape}
+          size="medium"
+          primaryColor={iconPrimaryColor}
+        />
       </DropdownLeftIconWrapper>
     );
     const iconAfter = (
       <DropdownRightIconWrapper>
-        <ChevronDownIcon label="chevron-icon" />
+        <ChevronDownIcon label="chevron-icon" primaryColor={iconPrimaryColor} />
       </DropdownRightIconWrapper>
     );