+++ packages/media/media-editor/src/react/editorView/toolbar/popups/colorButtonStyles.tsx
@@ -10,7 +10,9 @@ export const ColorSample: ComponentClass<HTMLAttributes<{}>> = styled.div`
   width: 24px;
   height: 24px;
   margin: 4px;
-  border-radius: ${borderRadius()};
+  border-radius: ${borderRadius()}px;
+  border-width: 2px;
+  border-style: solid;
 `;
 
 export const CheckArea: ComponentClass<HTMLAttributes<{}>> = styled.div`