+++ packages/monorepo-tooling/beautiful-insights/src/reports/duplicate-dependencies/duplicate-dependencies-report.ts
@@ -1,12 +1,14 @@
 import chalk from 'chalk';
 import crypto from 'crypto';
 import loadFileFromGitHistory from '../../util/load-file-from-git-history';
-import getDuplicateDependenciesReport from './get-duplicate-dependencies-report';
+import getDuplicateDependenciesReport, {
+  DuplicateDependencyReportEntry,
+} from './get-duplicate-dependencies-report';
 import Bitbucket, {
   CodeInsightsAnnotation,
   Severity,
   CodeInsightsReportResults,
-} from '../../reporters/bitbucket';
+} from '../../reporters/bitbucket-server';
 
 const REPORT_KEY = 'jira.frontend.duplicates';
 
@@ -40,18 +42,14 @@ function getTargetBranchYarnLockResolver(targetBranch: string) {
   return () => loadFileFromGitHistory(targetBranch, 'yarn.lock');
 }
 
-type RegressedDepenendency = {
+type RegressedDependencies = DuplicateDependencyReportEntry & {
   newVersions: string[];
-  name: string;
-  versions: string[];
-  isDevDependency: boolean;
 };
-export type RegressedDependencies = RegressedDepenendency[];
 
 async function getRegressedDependencies(
   sourceBranch: string,
   targetBranch: string,
-) {
+): Promise<RegressedDependencies[]> {
   const [currentBranch, masterDuplicatesReport] = await Promise.all([
     getDuplicateDependenciesReport({ returnAllDependencyCounts: true }),
     getDuplicateDependenciesReport({
@@ -61,7 +59,7 @@ async function getRegressedDependencies(
     }),
   ]);
 
-  const regressedDependencies: RegressedDependencies = currentBranch
+  const regressedDependencies = currentBranch
     .map(branchDependencyInfo => [
       masterDuplicatesReport.find(
         ({ name }) => branchDependencyInfo.name === name,
@@ -99,7 +97,7 @@ async function getRegressedDependencies(
       return {
         ...branchDependencyInfo,
         newVersions: branchDependencyInfo.versions.filter(
-          version => masterDependencyInfo.versions.indexOf(version) > -1,
+          version => masterDependencyInfo.versions.indexOf(version) === -1,
         ),
       };
     });
@@ -127,7 +125,7 @@ async function getRegressedDependencies(
 }
 
 async function publishInsightsReport(
-  regressedDependencies: RegressedDependencies,
+  regressedDependencies: RegressedDependencies[],
   bitbucket: Bitbucket,
 ) {
   // add base report
@@ -143,7 +141,7 @@ async function publishInsightsReport(
   );
 
   const annotations: CodeInsightsAnnotation[] = regressedDependencies.map(
-    ({ name, newVersions, isDevDependency }) => {
+    ({ name, newVersions, isDevDependency, directVersion }) => {
       const hash = crypto.createHash('sha256');
       hash.update(name, 'utf8');
 
@@ -154,7 +152,8 @@ async function publishInsightsReport(
         } extra versions of ${name} are introduced in this change`,
         path: 'yarn.lock',
         line: 0, // file level annotation
-        severity: isDevDependency ? Severity.LOW : Severity.HIGH,
+        severity:
+          !directVersion || isDevDependency ? Severity.LOW : Severity.HIGH,
       };
     },
   );
@@ -171,9 +170,9 @@ export default async function duplicateDependenciesReport(
   commit: string,
 ) {
   const bitbucket = new Bitbucket(gitUrl, token, commit);
-  //   const targetBranch =
-  //     (await bitbucket.getTargetBranch(sourceBranch)) || 'master';
-  const targetBranch = 'master';
+  const targetBranch =
+    (await bitbucket.getTargetBranch(sourceBranch)) || 'master';
+
   const regressedDependencies = await getRegressedDependencies(
     sourceBranch,
     targetBranch,
@@ -181,14 +180,3 @@ export default async function duplicateDependenciesReport(
 
   publishInsightsReport(regressedDependencies, bitbucket);
 }
-
-// interface Reporter {
-//   sendReport(): boolean;
-//   options: Object;
-// }
-
-// interface ReportOptions {
-//   reporters: Array<Reporter>;
-// }
-
-// module.exports = async function DuplicatesReport() {};