+++ packages/navigation/atlassian-switcher/src/providers/jira-data-providers.ts
@@ -7,4 +7,7 @@ export const MANAGE_HREF = '/plugins/servlet/customize-application-navigator';
 const fetchCustomLinks = () =>
   fetchJson<CustomLinksResponse>(`/rest/menu/latest/appswitcher`);
 
-export const CustomLinksProvider = asDataProvider(fetchCustomLinks);
+export const CustomLinksProvider = asDataProvider(
+  'customLinks',
+  fetchCustomLinks,
+);