+++ packages/media/media-picker/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/media-picker
 
+## 37.0.2
+- [patch] [b3c60e3c9c](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/b3c60e3c9c):
+
+  - Update media-editor dependency
+
 ## 37.0.1
 - [patch] [1bcaa1b991](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/1bcaa1b991):